requirements.txt
streamlit
pypdf
google-generativeai
plotly
python-dotenv
